backend/requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-environ>=0.11
psycopg2-binary>=2.9
celery[redis]>=5.3
redis>=5.0
django-storages[s3]>=1.14
boto3>=1.34
paho-mqtt>=1.6
django-encrypted-model-fields>=0.6
django-celery-beat>=2.6
Pillow>=10.0
requests>=2.31
jsonpath-ng>=1.6
cryptography>=41.0
